Line cleared bug corrected-new bug appeared
So, while doing the for loop to check every piece individually, one may delete an piece with no blocks. This will cause the for loop to skip an item, so the Piece.empty was created to signify a piece should be deleted without doing it right there.

Now, while deleting multiple lines, all the grounded pieces disappear. Further investigation is needed